+++ buffers/eros.pb.go
@@ -37,7 +37,7 @@ It has these top-level messages:
 */
 package protobufs
 
-import proto "github.com/golang/protobuf/proto"
+import proto "code.google.com/p/goprotobuf/proto"
 import math "math"
 
 // Reference imports to suppress errors if they are not otherwise used.