+++ util.go
@@ -140,3 +140,52 @@ func SendBroadcastAlert(predefined int32, message string) {
 
 	broadcastMessage("ALT", &bufmsg)
 }
+
+func ErosErrors(error_code int) error {
+	available_errors := map[int]string{
+		101: "Database read error",
+		102: "Database write error",
+		103: "Disk read error",
+		104: "Disk write error",
+		105: "Authentication error",
+		106: "Generic error",
+		107: "Bad name.",
+		108: "Name in use.",
+		109: "Cannot do that while matched via matchmaking.",
+		201: "Bad character info",
+		202: "Character already exists",
+		203: "Error while communicating with Battle.net",
+		204: "Verification failed.",
+
+		301: "Error processing replay",
+		302: "Error while processing match result",
+		303: "Duplicate Replay",
+		304: "The submitting client was not involved in the match.",
+		305: "Game too short.",
+		306: "Bad format. Required 1v1 with no observers.",
+		307: "Bad map. Require a map in the map pool.",
+		308: "All participants of the game must be registered.",
+		309: "Player not found in database.",
+		310: "You didn't play your matchmade opponent. You have been forfeited from that game.",
+		311: "The game was not played on Faster.",
+		312: "Cannot add veto. Map not in ranked pool.",
+		313: "Cannot add veto. Maximum number of vetoes used.",
+		314: "You are not in a game arranged by the Eros matchmaker.",
+		401: "Can't queue on this region without a character on this region.",
+		402: "The matchmaking request was cancelled.",
+		403: "Long process unavailable.",
+		501: "Chat room not joinable.",
+		502: "Bad password.",
+		503: "Can't create. Already exists.",
+		504: "Can't create. Room reserved.",
+		505: "Can't join. Max channel limit reached.",
+		506: "Can't send message. Not on channel.",
+		507: "Can't send message. User offline.",
+		508: "Can't send message. Missing fields.",
+		509: "Can't create room. Name too short.",
+		510: "Can't send message. Rate limit.",
+		511: "Can't send message. Message too long.",
+	}
+
+	return errors.New(available_errors[error_code])
+}